add Test.Arbitrary and refactor MnemonicsSpec
